Adds bitmap loading and some raster support for bitmaps.
Starts to flesh out the necessarly infrastructure to support bitmap
devices and drawing contexts.

Updates the heap allocator to allow mapping in existing memory so that
the Bitmap objects can share the same data that is already in system
memory.

The Heap object itself can be mapped into the Memory system and act like
a DataView.

The Memory implementation stores the segment and offset as properties
off of the mapped in DataView when it is allocated.

Adds MessageBox implementation (basic, not full featured) and more
advanced stubs for BitBlt.

Adds GetObject implementation for bitmaps.

Adds LoadBitmap implementation to load bitmaps from executable resource
tables.

Implements TextOut.

SkiFree is writing to an offscreen bitmap the animation frames for
certain things and it does not seem to position them correctly for some
reason. Not sure why, yet, but certainly lots of progress.

Adds the jsdoc and karma configurations.