Adds disk emulation and separates the DOS emulation.
Fixes some emulation issues with the i386 core mostly. I wish I had a
better way of testing it! I might rig up the dosbox core to run against
to test things.

Adds a FAT16 filesystem and the ability to add others. File I/O is
asynchronous, so parts of the emulator are reformed to allow for async
operations.

To that end, some kernel operations can be marked async (which means
they return a Promise) and if they return a Promise the task yields and
is only rescheduled with the Promise resolves. This includes file I/O
which should hopefully mean streaming CD content over a network (or, of
course, network content over the network) will do well.

Memory is now done via copy so that it can always be done synchronously.

A Stream class can cache network blocks and that is how we will handle
files. Even the items on the filesystem can be fake allocated but pull
data from the Stream so the filesystem can be realized without
downloading all the data. The data can then be pulled lazily.

I have learned that things use the Catch() function, the exception
handler function, to pull out the values of registers. How horrid!!
How deliciously horrendous!!! I love it. The chaos.

I'm excited for the midi stuff! There are some interesting soundfont
javascript libraries out there.